requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0
kafka-python==2.0.2
imapclient==2.3.1
email-reply-parser==0.5.12
